+++ harness/o.n.insane/src/org/netbeans/insane/impl/SmallObjectMap.java
@@ -60,14 +60,14 @@ class SmallObjectMap implements ObjectMap {
     public String getID(Object o) {
         // find whether it is known and wrapped
         Integer wid = wrappers.get(o);
-        if (wid != null) return getWrappedId(o, wid.intValue());
+        if (wid != null) return getWrappedId(o, wid);
 
         // ... or at least known
         if (isKnown(o)) return getNormalId(o);
 
         // unknown object
         if (putObject(o)) { //wrapped
-            return getWrappedId(o, wrappers.get(o).intValue());
+            return getWrappedId(o, wrappers.get(o));
         } else {
              return getNormalId(o);
         }
@@ -104,7 +104,7 @@ class SmallObjectMap implements ObjectMap {
         table[bucket] = o;
 
         // add the wrapping info
-        if (wrap) wrappers.put(o, Integer.valueOf(idCounter++));
+        if (wrap) wrappers.put(o, idCounter++);
         return wrap;
     }
